+++ src/main/java/com/breakinblocks/plonk/Plonk.java
@@ -12,7 +12,7 @@ import net.minecraftforge.fml.common.network.simpleimpl.SimpleNetworkWrapper;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
-@Mod(modid = Plonk.MOD_ID, name = Plonk.NAME, version = Plonk.VERSION)
+@Mod(modid = Plonk.MOD_ID, name = Plonk.NAME, version = Plonk.VERSION, guiFactory = "com.breakinblocks.plonk.client.gui.GuiFactoryPlonk")
 public class Plonk {
     public static final String MOD_ID = "plonk";
     public static final String NAME = "Plonk";