+++ src/main/java/com/breakinblocks/plonk/common/block/BlockPlacedItems.java
@@ -3,7 +3,6 @@ package com.breakinblocks.plonk.common.block;
 import com.breakinblocks.plonk.common.registry.RegistryMaterials;
 import com.breakinblocks.plonk.common.tile.TilePlacedItems;
 import com.breakinblocks.plonk.common.util.ItemUtils;
-import com.breakinblocks.plonk.common.util.bound.BoxCollection;
 import net.minecraft.block.Block;
 import net.minecraft.block.BlockChest;
 import net.minecraft.block.BlockDirectional;
@@ -175,7 +174,7 @@ public class BlockPlacedItems extends Block {
 
         if (worldId.rayTraceBlocks(from, to, false, false, true) != null) {
             TilePlacedItems tile = (TilePlacedItems) worldId.getTileEntity(pos);
-            int index = tile.getContentsBoxes().selectionLastEntry.id;
+            int index = tile.getContentsBoxes().getSelectedId();
             int slot = index <= 0 ? 0 : index - 1;
             ItemStack stack = tile.getStackInSlot(slot);
             if (!stack.isEmpty()) {
@@ -233,9 +232,7 @@ public class BlockPlacedItems extends Block {
     public ItemStack getPickBlock(IBlockState state, RayTraceResult target, World world, BlockPos pos, EntityPlayer player) {
         TilePlacedItems tile = (TilePlacedItems) world.getTileEntity(pos);
         if (tile == null) return ItemStack.EMPTY;
-        BoxCollection.Entry last = tile.getContentsBoxes().selectionLastEntry;
-        if (last == null) return ItemStack.EMPTY;
-        int index = last.id;
+        int index = tile.getContentsBoxes().getSelectedId();
         int slot = index <= 0 ? 0 : index - 1;
         return tile.getStackInSlot(slot);
     }