+++ src/main/java/com/breakinblocks/plonk/common/util/bound/BoxCollection.java
@@ -19,8 +19,8 @@ public class BoxCollection {
     private final ArrayList<Entry> collisionBoxes;
     private final ArrayList<Entry> selectionBoxes;
     private final Box renderBox;
-    public Entry selectionLastEntry = null;
-    public AxisAlignedBB selectionLastAABB = null;
+    private Entry selectionLastEntry = null;
+    private AxisAlignedBB selectionLastAABB = null;
     private Entry boundsEntry = null;
 
     private BoxCollection(ArrayList<Entry> boxes) {
@@ -70,6 +70,15 @@ public class BoxCollection {
         if (selectionLastAABB == null)
             return renderBox.toAABB();
         return selectionLastAABB;
+    }
+
+    /**
+     * Gets the last selected id if available
+     *
+     * @return returns the last selected id or -1 if none
+     */
+    public int getSelectedId() {
+        return selectionLastEntry == null ? -1 : selectionLastEntry.id;
     }
 
     public RayTraceResult collisionRayTrace(Block block, ICollisionRayTrace collisionRayTrace, IBlockState blockState, World worldIn, BlockPos pos, Vec3d start, Vec3d end) {